requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
paho-mqtt>=1.6.0
confluent-kafka>=2.3.0
protobuf>=4.25.0
grpcio-tools>=1.60.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
